Agregado: Pruebas con Onda real del LAT. Correciones menores de nombres de funciones.

+++ src/Administrar_Carpetas.py
@@ -145,6 +145,7 @@ class FileManager:
 
             for line in f:
                 if line.strip():
-                    data.append(int(line.strip()))
+                    #data.append(int(line.strip()))
+                    data.append(float(line.strip()))
 
         return metadata, data
